Participants feature added

+++ myapp/views.py
@@ -70,6 +70,7 @@ def home(request):
 def room(request, pk):
     room = Room.objects.get(id=pk)
     room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
 
     if request.method == 'POST':
         message = Message.objects.create(
@@ -77,9 +78,10 @@ def room(request, pk):
             room=room,
             body=request.POST.get('body')
         )
+        room.participants.add(request.user)
         return redirect('room', pk=room.id)
 
-    context = {'room' : room, 'room_messages' : room_messages}
+    context = {'room' : room, 'room_messages' : room_messages, 'participants' : participants}
     return render(request, 'myapp/room.html', context)
 
 @login_required(login_url='login')
@@ -131,7 +133,23 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed Here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'myapp/delete.html', {'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowed Here!!')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'myapp/delete.html', {'obj':message})
